Revert "plugin rename"
