+++ scripts/build.sh
@@ -3,12 +3,12 @@
 GITVERSION=`git describe --always`
 SOURCEDIR=$1
 BUILDDIR=$SOURCEDIR/build
-PLUGIN=`echo $SOURCEDIR | grep -oh "snap-.*"`
+PLUGIN=`echo $SOURCEDIR | grep -oh "pulse-.*"`
 ROOTFS=$BUILDDIR/rootfs
 BUILDCMD='go build -a -ldflags "-w"'
 
 echo
-echo "****  Snap Plugin Build  ****"
+echo "****  Pulse Plugin Build  ****"
 echo
 
 # Disable CGO for builds
@@ -22,5 +22,5 @@ mkdir -p $ROOTFS
 
 # Build plugin
 echo "Source Dir = $SOURCEDIR"
-echo "Building Snap Plugin: $PLUGIN"
+echo "Building Pulse Plugin: $PLUGIN"
 $BUILDCMD -o $ROOTFS/$PLUGIN
